* feature/redis:
  [손원철]<chore>: 사용하지 않는 클래스 삭제.
  [손원철]<fix>: 임박한 나의 재료 조회 :  responose 값에 Mark_name  추가              <fix> : 나의 storage별 조회 :  유통기한만료 및 당일 재료 respone값 수정              <update> : token이 만료될시 응답 메시지 추가
  [손원철]<update,fix>  Redis (리프레쉬토큰, 나의 냉장고 조회, 오늘 할 요리 조회) 캐싱 기능구현 및 재료 Mark_name 열 추가

+++ src/main/java/com/sparta/cookbank/service/DoneRecipeService.java
@@ -8,6 +8,7 @@ import com.sparta.cookbank.domain.recipe.Recipe;
 import com.sparta.cookbank.domain.myingredients.MyIngredients;
 import com.sparta.cookbank.domain.recipe.dto.RecipeFixRequestDto;
 import com.sparta.cookbank.domain.recipe.dto.RecipeFixResponseDto;
+import com.sparta.cookbank.redis.ingredient.RedisIngredientRepo;
 import com.sparta.cookbank.repository.DoneRecipeRepository;
 import com.sparta.cookbank.repository.MemberRepository;
 import com.sparta.cookbank.repository.MyIngredientsRepository;
@@ -31,24 +32,30 @@ public class DoneRecipeService {
     private final RecipeRepository recipeRepository;
     private final MyIngredientsRepository myIngredientsRepository;
     private final DoneRecipeRepository doneRecipeRepository;
+    private final RedisIngredientRepo redisIngredientRepo;
 
 
     public void UsedIngredient(Long recipeId, DoneRecipeRequestDto requestDto) {
-        for (Long id : requestDto.getIngredients_id()) {
-            MyIngredients ingredients = myIngredientsRepository.findById(id).orElseThrow(
-                    () -> new IllegalArgumentException("해당 재료가 없습니다.")
-            );
-            myIngredientsRepository.delete(ingredients);
+        if(!requestDto.getIngredients_id().isEmpty()){
+            for (Long id : requestDto.getIngredients_id()) {
+                MyIngredients ingredients = myIngredientsRepository.findById(id).orElseThrow(
+                        () -> new IllegalArgumentException("해당 재료가 없습니다.")
+                );
+                myIngredientsRepository.delete(ingredients);
+                // 레디스 캐시 초기화.
+                redisIngredientRepo.deleteAll();
+        }}
 
-            Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-                    () -> new IllegalArgumentException("유저정보가 올바르지 않습니다.")
-            );
-            Recipe recipe = recipeRepository.findById(recipeId).orElseThrow(
-                    () -> new IllegalArgumentException("해당 레시피가 존재하지 않습니다.")
-            );
-            DoneRecipe doneRecipe = new DoneRecipe(member, recipe);
-            doneRecipeRepository.save(doneRecipe);
-        }
+
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
+                () -> new IllegalArgumentException("유저정보가 올바르지 않습니다.")
+        );
+        Recipe recipe = recipeRepository.findById(recipeId).orElseThrow(
+                () -> new IllegalArgumentException("해당 레시피가 존재하지 않습니다.")
+        );
+        DoneRecipe doneRecipe = new DoneRecipe(member, recipe);
+        doneRecipeRepository.save(doneRecipe);
+
     }
